+++ src/app/_home/components/footer.mdx
