+++ src/app/writing/components/writing-page.tsx
