+++ src/app/writing/mountain-tracking/page.mdx
@@ -1,5 +1,6 @@
-import WritingPage from '@/components/writing-page'
 import { buildArticleMetadata } from '@/lib/metadata'
+
+import WritingPage from '../components/writing-page'
 
 export default function Page({ children }) {
     return <WritingPage meta={meta}>{children}</WritingPage>
