--- src/components/home-page/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './home-page'
